Rewrite in C++
Rewrote the plugin in C++ so it works with Compiz 0.9.x.  This is
more-or-less complete and seems to work reasonably well excepting that
the filter expression for windows to consider is inordinately complex,
and that it has the same bug the old plugin had that it can get stuck
on a single set of windows in the same geometry.

A solution to that problem should be relatively simple.