Handle collisions, a bit.
Handle collisions with zero distance between the top left corners (in
the appropriate direction) by chosing the smaller or larger of the two
windows, and ignoring complete pixel-per-pixel collisions, but
recording them in an array within the search object.

This could probably be better handled by simply making our
calculations based on the centerpoint of the window.