Use centerpoints for focus.
Finally wrapping up this version of the plugin, replaced the logic to
fallback to width and handle collision detection with simply the use
of centerpoints in the distance calcuation.

This probably might not work very well on larger displays, where
windows are stacked in a tiled fashion, since "down" might not always
reflect what the user was thinking.  We can solve that by adding
lateral collision to the candidate filter.  It seems to otherwise
perfectly enable rather reasonable behaviour.